# api/docs/telephant-api-documentation.yml
openapi: 3.0.0
info:
  title: Telephant API
  description: |
    *The mechanism for swapping airtime is powered by Telephant's APIs. Developers can use the API to tap into the platform's capabilities and incorporate the service into their own apps.*

    The Telephant API's backbone is HTTP. The nature of our API is RESTful. It has a set of predictable resource URLs. It returns HTTP response codes to indicate errors. It also accepts JSON as an HTTP body payload and returns it, along with some other useful formats.
    
    You can use your favorite HTTP/REST library for your programming language or an HTTP client like Postman or CURL to use Telephant's API.

    Happy programming **:-)**

    ## HTTP Response Status Codes
    HTTP response status codes show whether or not a particular HTTP request was completed successfully. Section 10 of RFC 2616 defines the status codes listed below. RFC 7231 contains an updated specification.

    Telephant API makes use of the following HTTP status codes, grouped into three classes:

    ### Successful Responses - 2XX

    #### 200 - OK

    The request has succeeded.

    #### 201 - Created

    As a result of the request's success, a new resource has been generated. This is the response that is normally sent in response to POST or some PUT queries.

    ### Client Error - 4XX
    
    #### 400 - Bad Request
    Due to incorrect syntax, the server was unable to understand the request.
    
    #### 401 - Unauthorized
...
    
...
    
...
    
...

...

...

...

...

...
servers:
...
paths:
...
  /v1/contacts:

    get:
      summary: List the contacts.
      description: List the contacts of an individual or group account. 
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
        - in: query
          name: page
          description: The page number of results to return.
          schema:
              type: integer
        - in: query
          name: limit
          description: The numbers of items to return per page.
          schema:
            type: integer
      responses:
        '2XX':
          description: Success
          content:
            application/json:
              examples:
                Contacts-Get:
                  $ref: '#/components/examples/Contacts-Get'
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'

    post:
      summary: Create contact.
      description: |
        Create a contact for a individual or group account.

        **Note** The account or username that makes the request will be assigned as the owner account.  It is not possible to change a contact's ownership to multiple owners.
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
      requestBody:
        description: |
          Define the properties of the contact to create.

          You can set a property to null.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - telephoneNumber
              properties:
                telephoneNumber:
                  type: string
                  pattern: '^(27)[0-9]{9}$'
                  description: Provide a valid South African mobile telphone number in the format 27XXXXXXXXX.
                  example: "27821234567"
                  required: true
                firstName:
                  description: You can only use alpha characters in a name.
                  type: string
                  maxLength: 30
                  example: Piet
                  pattern: '^[a-zA-Z]*$'
                lastName:
                  description: You can only use alpha characters in a name.
                  type: string
                  maxLength: 30
                  example: Pompies
                  pattern: '^[a-zA-Z]*$'
                email:
                  description: A valid email address for the contact.
                  type: string
                  format: email
                  example: piet@example.com
      responses: 
        '200':
          description: Success
          content:
            application/json:
              example:
                { 
                  message: 'The contact was successfully created.' 
                }   
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'
  /v1/contacts/{contact_id}:
    get:
      summary: View specific contact.
      description: Get the details of a specific contact from an individual or group account.
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: contact_id
          schema:
            type: string
            format: uuid
          required: true
          description: |
            A contact's unique ID from the contacts list.

            Because it is possible to have two contacts with the same phone number in your contacts list, the contacts _id must be used to get, update or delete the contact from the list.
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '200':
          description: Success
          content: 
            application/json:
              examples:
                Contact-Get:
                  $ref: '#/components/examples/Contact-Get'
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'

    put:
      summary: Update a specific contact.
      description: Update a specific contact for an individual or group account.
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: contact_id
          schema:
            type: string
            format: uuid
          required: true
          description: |
            A contact's unique ID from the contacts list.

            Because it is possible to have two contacts with the same phone number in your contacts list, the contacts id must be used to get, update or delete the contact from the list.

        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
      requestBody:
        description: |
          Define the properties of the contact to update.

          You can set a property to null.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - telephoneNumber
              properties:
                telephoneNumber:
                  type: string
                  pattern: '^(27)[0-9]{9}$'
                  description: Provide a valid South African mobile telphone number in the format 27XXXXXXXXX.
                  example: "27821234567"
                  required: true
                firstName:
                  description: You can only use alpha characters in a name.
                  type: string
                  maxLength: 30
                  example: Piet
                  pattern: '^[a-zA-Z]*$'
                lastName:
                  description: You can only use alpha characters in a name.
                  type: string
                  maxLength: 30
                  example: Pompies
                  pattern: '^[a-zA-Z]*$'
                email:
                  description: A valid email address for the contact.
                  type: string
                  format: email
                  example: piet@example.com
      responses:
        '200':
          description: Success
          content: 
            application/json:
              example:
                {
                    message: 'The contact was successfully updated.'
                }
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'

    delete:
      summary: Delete specific contact.
      description: Delete a specific contact for an individual or group account.
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: contact_id
          schema:
            type: string
            format: uuid
          required: true
          description: |
            A contact's unique ID from the contacts list.

            Because it is possible to have two contacts with the same phone number in your contacts list, the contacts _id must be used to get, update or delete the contact from the list.

        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '200':
          description: Success
          content: 
            application/json:
              example:
                {
                    message: 'Contact succesfully deleted.'
                }
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'

#### NOTIFICATIONS ####
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    4XX:
      description: |
        Client error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: Integer
                example: 400
                description: |
                  Please review the documentation for [HTTP Response Status Codes](/#section/HTTP-Response-Status-Codes)

                  Possible values are:
                    - 400 - Bad Request
                    - 401 - Unauthorized
                    - 403 - Forbidden
                    - 404 - Not Found
                    - 405 - Method Not Allowed
              name:
                type: String
                example: 'BadRequest'
                description: This value will always be returned as 'bumpy case'.
              message:
                type: String
                description: |
                  The message field will provide specific information as to the reason for the error.
                example: 'The account ID provided is not valid.'
    5XX:
      description: |
        Server error  
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: Integer
                example: 500
              name:
                type: String
                example: 'InternalServerError'
              message:
                type: String
                example: 'We are sorry but something went wrong on our side.'
  examples:
...
    Contact-Get:
      value:
        [
            {
                "_id": "610557d603e71002e0219158",
                "telephoneNumber": "27434921801",
                "firstName": null,
                "lastName": "Gatlin",
                "owner": "27829524031",
                "__v": 0
            }
        ]
    Contacts-Get:
      value:
        {
    "page": 1,
    "total_pages": 5,
    "total_results": 41,
    "results": [
        {
            "owner": [
                "27829524031"
            ],
            "_id": "61027161e2eb5f1e6831ea09",
            "telephoneNumber": "27828400904",
            "firstName": "Avrom",
            "lastName": "Ateggart",
            "email": "aateggart4@linkedin.com"
        },
        {
            "owner": [
                "27829524031"
            ],
            "_id": "61027161e2eb5f1e6831ea0e",
            "telephoneNumber": "27962371963",
            "firstName": "Berkie",
            "lastName": null
        },
        {
            "owner": [
                "27829524031"
            ],
            "_id": "61027161e2eb5f1e6831ea13",
            "telephoneNumber": "27415630343",
            "firstName": "Hyman",
            "lastName": "Maudlin",
            "email": "hmaudline@huffingtonpost.com"
        },
        {
            "owner": [
                "27829524031"
            ],
            "_id": "61027161e2eb5f1e6831ea18",
            "telephoneNumber": "27422531972",
            "firstName": "Ignatius",
            "lastName": "Bransdon"
        },
        {
            "owner": [
                "27829524031"
            ],
            "_id": "61027161e2eb5f1e6831ea1d",
            "telephoneNumber": "27534405808",
            "firstName": "Donica",
            "lastName": "Ussher"
        },
        {
            "owner": [
                "27829524031"
            ],
            "_id": "61027161e2eb5f1e6831ea22",
            "telephoneNumber": "27706368969",
            "firstName": "Dreddy",
            "lastName": null
        },
        {
            "owner": [
                "27829524031"
            ],
            "_id": "61027161e2eb5f1e6831ea27",
            "telephoneNumber": "27691370795",
            "firstName": "Jeremie",
            "lastName": "Desport"
        },
        {
            "owner": [
                "27829524031"
            ],
            "_id": "61027161e2eb5f1e6831ea05",
            "telephoneNumber": "27847470159",
            "firstName": "Bartlet",
            "lastName": null,
            "email": "bfairbourne0@nymag.com"
        },
        {
            "owner": [
                "27829524031"
            ],
            "_id": "61027161e2eb5f1e6831ea0a",
            "telephoneNumber": "27230295211",
            "firstName": "Bronnie",
            "lastName": "Meader"
        },
        {
            "owner": [
                "27829524031"
            ],
            "_id": "61027161e2eb5f1e6831ea0f",
            "telephoneNumber": "27129256706",
            "firstName": null,
            "lastName": "Giamitti",
            "email": "agiamittia@bluehost.com"
        }
    ]
}
...